# packages/python/kafka-messaging-internal/contracts/openapi/v1.yaml
openapi: 3.0.3
info:
  title: Kafka Messaging Internal API
  description: REST API for messaging operations including database, schema registry, and event handling
  version: 1.0.0
  contact:
    name: LLM Observability Platform
    email: dev@example.com
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
paths:
...
  /api/v1/schema-registry/schemas:
    post:
      summary: Register new schema
      description: Registers a new schema in the schema registry
      operationId: registerSchema
      tags:
        - Schema Registry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaRegisterRequest'
      responses:
        '201':
          description: Schema registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  schema_id:
                    type: integer
                    example: 123
                  subject:
                    type: string
                    example: "user-events"
                  version:
                    type: integer
                    example: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      summary: List all subjects
      description: Retrieves all available schema subjects
      operationId: listSubjects
      tags:
        - Schema Registry
      responses:
        '200':
          description: Subjects retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  subjects:
                    type: array
                    items:
                      type: string
                    example: ["user-events", "order-events"]
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/schema-registry/schemas/{subject}:
    get:
      summary: Get schema by subject
      description: Retrieves schema information for a specific subject
      operationId: getSchemaBySubject
      tags:
        - Schema Registry
      parameters:
        - name: subject
          in: path
          required: true
          schema:
            type: string
          description: Schema subject name
        - name: version
          in: query
          required: false
          schema:
            type: integer
          description: Specific version to retrieve (latest if not specified)
      responses:
        '200':
          description: Schema retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaInfoResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/schema-registry/compatibility:
    post:
      summary: Check schema compatibility
      description: Checks if a schema is compatible with existing versions
      operationId: checkCompatibility
      tags:
        - Schema Registry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaCompatibilityRequest'
      responses:
        '200':
          description: Compatibility check completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  is_compatible:
                    type: boolean
                    example: true
                  compatibility_level:
                    type: string
                    example: "BACKWARD"
                  message:
                    type: string
                    example: "Schema is compatible"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/schema-registry/compatibility/{subject}:
    put:
      summary: Update compatibility level
      description: Updates the compatibility level for a subject
      operationId: updateCompatibility
      tags:
        - Schema Registry
      parameters:
        - name: subject
          in: path
          required: true
          schema:
            type: string
          description: Schema subject name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompatibilityUpdateRequest'
      responses:
        '200':
          description: Compatibility updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  subject:
                    type: string
                    example: "user-events"
                  compatibility:
                    type: string
                    example: "BACKWARD"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/schema-registry/serialize:
    post:
      summary: Serialize data
      description: Serializes data using a registered schema
      operationId: serializeData
      tags:
        - Schema Registry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SerializationRequest'
      responses:
        '200':
          description: Data serialized successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SerializationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/schema-registry/deserialize:
    post:
      summary: Deserialize data
      description: Deserializes data using a registered schema
      operationId: deserializeData
      tags:
        - Schema Registry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeserializationRequest'
      responses:
        '200':
          description: Data deserialized successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    additionalProperties: true
                  schema_id:
                    type: integer
                    example: 123
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    SchemaRegisterRequest:
      type: object
      properties:
        subject:
          type: string
          description: Schema subject name
          example: "user-events"
        schema:
          type: string
          description: Schema definition (AVRO/JSON/PROTOBUF)
          example: '{"type":"record","name":"UserEvent","fields":[{"name":"user_id","type":"string"},{"name":"action","type":"string"}]}'
        schema_type:
          $ref: '#/components/schemas/SchemaType'
      required:
        - subject
        - schema
        - schema_type
    SchemaType:
      type: string
      enum:
        - AVRO
        - JSON
        - PROTOBUF
      description: Schema type
    SchemaInfoResponse:
      type: object
      properties:
        subject:
          type: string
          description: Schema subject name
          example: "user-events"
        schema_id:
          type: integer
          description: Schema ID
          example: 123
        schema_type:
          type: string
          description: Schema type
          example: "AVRO"
        schema:
          type: string
          description: Schema definition
          example: '{"type":"record","name":"UserEvent","fields":[{"name":"user_id","type":"string"},{"name":"action","type":"string"}]}'
        version:
          type: integer
          description: Schema version
          example: 1
        compatibility:
          type: string
          description: Compatibility level
          example: "BACKWARD"
          nullable: true
    SchemaCompatibilityRequest:
      type: object
      properties:
        subject:
          type: string
          description: Schema subject name
          example: "user-events"
        schema:
          type: string
          description: Schema definition to check compatibility
          example: '{"type":"record","name":"UserEvent","fields":[{"name":"user_id","type":"string"},{"name":"action","type":"string"},{"name":"timestamp","type":"long"}]}'
        schema_type:
          $ref: '#/components/schemas/SchemaType'
      required:
        - subject
        - schema
        - schema_type
    CompatibilityUpdateRequest:
      type: object
      properties:
        subject:
          type: string
          description: Subject to update compatibility for
          example: "user-events"
        compatibility:
          type: string
          description: Compatibility level
          enum:
            - BACKWARD
            - FORWARD
            - FULL
            - NONE
          example: "BACKWARD"
      required:
        - subject
        - compatibility
    SerializationRequest:
      type: object
      properties:
        subject:
          type: string
          description: Subject for serialization
          example: "user-events"
        data:
          description: Data to serialize
          example:
            user_id: "user-123"
            action: "login"
        schema_id:
          type: integer
          description: Optional schema ID to use
          example: 123
          nullable: true
      required:
        - subject
        - data
    SerializationResponse:
      type: object
      properties:
        data:
          type: string
          description: Serialized data (base64 encoded)
          example: "eyJ1c2VyX2lkIjoidXNlci0xMjMiLCJhY3Rpb24iOiJsb2dpbiJ9"
        schema_id:
          type: integer
          description: Schema ID used for serialization
          example: 123
          nullable: true
    DeserializationRequest:
      type: object
      properties:
        data:
          type: string
          description: Base64-encoded data to deserialize
          example: "eyJ1c2VyX2lkIjoidXNlci0xMjMiLCJhY3Rpb24iOiJsb2dpbiJ9"
        schema_id:
          type: integer
          description: Schema ID to use for deserialization
          example: 123
      required:
        - data
        - schema_id
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request format"
        details:
          type: object
          additionalProperties: true
          description: Additional error details
          example:
            field: "topic"
            reason: "Topic name is required"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2023-01-01T12:00:00Z"
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []